Create Articulate_Points_Bridges.py
Articulation Point: In a graph, a vertex is called an articulation point if removing it and all the edges associated with it results in the increase of the number of connected components in the graph.The complexity of the graph is O(V+E) V vertices E Edges
Bridge: an edge in a graph between vertices say u  and vis called a Bridge, if after removing it, there will be no path left between and . It's definition is very similar to that of Articulation Points. Just like them it also represents vulnerabilities in the given network.